+++ GameModeManager/Features/RTVCommands.cs
@@ -1,15 +1,8 @@
 // Included libraries
-using WASDSharedAPI;
 using GameModeManager.Core;
-using GameModeManager.Menus;
-using CounterStrikeSharp.API;
-using GameModeManager.Models;
 using GameModeManager.Contracts;
 using CounterStrikeSharp.API.Core;
 using Microsoft.Extensions.Logging;
-using GameModeManager.CrossCutting;
-using Microsoft.Extensions.Localization;
-using CounterStrikeSharp.API.Modules.Menu;
 using CounterStrikeSharp.API.Modules.Admin;
 using CounterStrikeSharp.API.Modules.Commands;
 
@@ -21,33 +14,19 @@ namespace GameModeManager.Features
     {
         // Define class dependencies
         private Plugin? _plugin;
-        private RTVMenus _rtvMenus;
-        private GameRules _gameRules;
         private RTVManager _rtvManager;
         private PluginState _pluginState;
-        private VoteManager _voteManager;
-        private MenuFactory _menuFactory;
-        private StringLocalizer _localizer;
         private Config _config = new Config();
         private ILogger<ModeCommands> _logger;
-        private MaxRoundsManager _maxRoundsManager;
-        private TimeLimitManager _timeLimitManager;
-        private VoteOptionManager _voteOptionManager;
+        private AsyncVoteManager _asyncVoteManager;
 
         // Define class instance
-        public RTVCommands(PluginState pluginState, IStringLocalizer iLocalizer, ILogger<ModeCommands> logger, RTVManager rtvManager, GameRules gameRules, VoteManager voteManager, MaxRoundsManager maxRoundsManager, TimeLimitManager timeLimitManager, VoteOptionManager voteOptionManager, RTVMenus rtvMenus, MenuFactory menuFactory)
+        public RTVCommands(PluginState pluginState, ILogger<ModeCommands> logger, RTVManager rtvManager, AsyncVoteManager asyncVoteManager)
         {
             _logger = logger;
-            _rtvMenus = rtvMenus;
-            _gameRules = gameRules;
             _rtvManager = rtvManager;
-            _voteManager = voteManager;
             _pluginState = pluginState;
-            _menuFactory = menuFactory;
-            _maxRoundsManager = maxRoundsManager;
-            _timeLimitManager = timeLimitManager; 
-            _voteOptionManager = voteOptionManager;
-            _localizer = new StringLocalizer(iLocalizer, "rtv.prefix");
+            _asyncVoteManager = asyncVoteManager;
         }
 
         // Load config
@@ -75,53 +54,7 @@ namespace GameModeManager.Features
             }
         }
 
-        // Define class methods
-        private void StartVote(VoteResult? result, CCSPlayerController? player)
-        {
-            if(player?.PlayerName != null && result?.VoteCount != null)
-            {
-                // Display to chat
-                Server.PrintToChatAll($"{_localizer.LocalizeWithPrefix("rtv.rocked-the-vote", player.PlayerName)} {_localizer.Localize("general.votes-needed", result.VoteCount, result.RequiredVotes)}");
-                Server.PrintToChatAll(_localizer.LocalizeWithPrefix("rtv.votes-reached"));
-            }
-
-            // Load Options
-            _voteOptionManager.LoadOptions();
-            _rtvMenus.Load(_voteOptionManager.ScrambleOptions());
-
-            // Start vote
-            _voteManager.StartVote(_pluginState.RTVDuration);
-
-            // Display vote menu
-            if (_config.RTV.Style.Equals("wasd", StringComparison.OrdinalIgnoreCase))
-            {
-                foreach (var validPlayer in Extensions.ValidPlayers())
-                {
-                    IWasdMenu? menu;
-                    menu = _rtvMenus.GetWasdMenu();
-
-                    if (menu != null)
-                    {
-                        _menuFactory.OpenWasdMenu(validPlayer, menu);
-                    }
-                }
-            }
-            else
-            {
-                foreach (var validPlayer in Extensions.ValidPlayers())
-                {
-                    BaseMenu menu;
-                    menu = _rtvMenus.GetMenu();
-
-                    if (menu != null)
-                    {
-                        _menuFactory.OpenMenu(menu, validPlayer);
-                    }
-                }
-            }
-        }
-
-        // Define client rtv duration command handler
+        // Define command handlers
         [RequiresPermissions("@css/cvar")]
         [CommandHelper(minArgs: 1, usage: "<duration>", whoCanExecute: CommandUsage.SERVER_ONLY)]
         public void OnRTVDurationCommand(CCSPlayerController? player, CommandInfo command)
@@ -136,7 +69,6 @@ namespace GameModeManager.Features
             return;
         }
 
-        // Define client rtv seconds before end command handler
         [RequiresPermissions("@css/cvar")]
         [CommandHelper(minArgs: 1, usage: "<seconds>",whoCanExecute: CommandUsage.SERVER_ONLY)]
         public void OnRTVSecondsBeforeEndCommand(CCSPlayerController? player, CommandInfo command)
@@ -151,7 +83,6 @@ namespace GameModeManager.Features
             return;
         }
 
-        // Define client rtv rounds before end command handler
         [RequiresPermissions("@css/cvar")]
         [CommandHelper(minArgs: 1, usage: "<rounds>", whoCanExecute: CommandUsage.SERVER_ONLY)]
         public void OnRTVRoundsBeforeEndCommand(CCSPlayerController? player, CommandInfo command)
@@ -166,7 +97,6 @@ namespace GameModeManager.Features
             return;
         }
 
-        // Define server rtv end of map command handler
         [RequiresPermissions("@css/cvar")]
         [CommandHelper(minArgs: 1, usage: "<true|false>", whoCanExecute: CommandUsage.SERVER_ONLY)]
         public void OnRTVEndOfMapVoteCommand(CCSPlayerController? player, CommandInfo command)
@@ -181,7 +111,6 @@ namespace GameModeManager.Features
             return;
         }
 
-        // Define server rtv start vote command handler
         [RequiresPermissions("@css/cvar")]
         [CommandHelper(minArgs: 2, usage: "<duration> <true|false>", whoCanExecute: CommandUsage.SERVER_ONLY)]
         public void OnRTVStartVoteCommand(CCSPlayerController? player, CommandInfo command)
@@ -197,77 +126,22 @@ namespace GameModeManager.Features
                 {
                     _pluginState.ChangeImmediately = changeImmediately;
                 }
-                StartVote(null, null);
+                _asyncVoteManager.StartVote(null, null);
             }
             return;
         }
 
-        // Define client rtv command handler
         [RequiresPermissions("@css/cvar")]
         [CommandHelper(whoCanExecute: CommandUsage.CLIENT_ONLY)]
         public void OnRTVCommand(CCSPlayerController? player, CommandInfo command)
         {
             if (player != null)
             {
-                if (_pluginState.EofVoteHappened)
-                {
-                    player.PrintToChat(_localizer.LocalizeWithPrefix("rtv.schedule-change"));
-                    return;
-                }
-
-                if (_pluginState.DisableCommands || !_config.RTV.Enabled)
-                {
-                    player.PrintToChat(_localizer.LocalizeWithPrefix("general.validation.disabled"));
-                    return;
-                }
-
-                if (_gameRules.WarmupRunning)
-                {
-                    if (!_config.RTV.EnabledInWarmup)
-                    {
-                        player.PrintToChat(_localizer.LocalizeWithPrefix("general.validation.warmup"));
-                        return;
-                    }
-                }
-                else if (_timeLimitManager.UnlimitedTime() && !_maxRoundsManager.UnlimitedRounds && _config.RTV.MinRounds > 0 && _config.RTV.MinRounds > _gameRules.TotalRoundsPlayed)
-                {
-                    player!.PrintToChat(_localizer.LocalizeWithPrefix("general.validation.minimum-rounds", _config.RTV.MinRounds));
-                    return;
-                }
-
-                if (Extensions.ValidPlayerCount() < _config!.RTV.MinPlayers)
-                {
-                    player.PrintToChat(_localizer.LocalizeWithPrefix("general.validation.minimum-players", _config!.RTV.MinPlayers));
-                    return;
-                }
-
-                if (_pluginState.NextMap != null & _pluginState.NextMap != null)
-                {
-                    player.PrintToChat(_localizer.LocalizeWithPrefix("general.validation.disabled"));
-                    return;
-                }
-
-                VoteResult result = _rtvManager.AddVote(player.UserId!.Value);
-                switch (result.Result)
-                {
-                    case VoteResultEnum.Added:
-                        Server.PrintToChatAll($"{_localizer.LocalizeWithPrefix("rtv.rocked-the-vote", player.PlayerName)} {_localizer.Localize("general.votes-needed", result.VoteCount, result.RequiredVotes)}");
-                        break;
-                    case VoteResultEnum.AlreadyAddedBefore:
-                        player.PrintToChat($"{_localizer.LocalizeWithPrefix("rtv.already-rocked-the-vote")} {_localizer.Localize("general.votes-needed", result.VoteCount, result.RequiredVotes)}");
-                        break;
-                    case VoteResultEnum.VotesAlreadyReached:
-                        player.PrintToChat(_localizer.LocalizeWithPrefix("rtv.disabled"));
-                        break;
-                    case VoteResultEnum.VotesReached:
-                        StartVote(result, player);
-                        break;
-                }
+                _asyncVoteManager.RTVCounter(player);
             }
             return;
         }
 
-        // Define server rtv command handler
         [CommandHelper(minArgs: 1, usage: "<true|false>", whoCanExecute: CommandUsage.SERVER_ONLY)]
         public void OnRTVEnabledCommand(CCSPlayerController? player, CommandInfo command)
         {
@@ -296,13 +170,13 @@ namespace GameModeManager.Features
             }
         }
 
-        // Define player disconnected event handler
+        // Define event handler
         public HookResult PlayerDisconnected(EventPlayerDisconnect @event, GameEventInfo @eventInfo)
         {
             var player = @event.Userid;
             if (player?.UserId != null)
             {
-                _rtvManager!.RemoveVote(player.UserId.Value);
+                _asyncVoteManager!.RemoveVote(player.UserId.Value);
             }
             return HookResult.Continue;
         }
